pyproject.toml
[tool.poetry]
name = "core-pptautomate"
version = "0.9.0"
description = "PPT file automation for Market-Report"
authors = ["en_kanoe <en_kanoe@outlook.jp>"]
readme = "README.md"
packages = [
    {include = "api", from = "src"},
    {include = "services", from = "src"},
    {include = "tests", from = "src"},
]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "0.115.4"
gunicorn = "23.0.0"
langchain-core = "0.2.5"
langchain-openai = "0.1.8"
langchain = "0.2.3"
openai = "1.55.3"
python-dotenv = "1.0.1"
python-multipart = "0.0.16"
python-pptx = "1.0.2"
uvicorn = "0.32.0"
pptx-ea-font = "0.0.4"
azure-monitor-opentelemetry = "1.2.0"
azure-monitor-opentelemetry-exporter = "^1.0.0b32"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
